% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{animate_eq_li}
\alias{animate_eq_li}
\title{Use ggplotly to show an animation of the payoff sets of a list of equilibria}
\usage{
animate_eq_li(g, eq.li, x = g$sdf$x[1], ...)
}
\description{
Use ggplotly to show an animation of the payoff sets of a list of equilibria
}
