% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_eq.R
\name{compare_eq}
\alias{compare_eq}
\title{Helper function to find differences between two equilibria}
\usage{
compare_eq(eq1, eq2 = g[["eq"]], g, verbose = TRUE)
}
\description{
Helper function to find differences between two equilibria
}
