% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{diagnose_transitions}
\alias{diagnose_transitions}
\title{Take a look at the computed transitions for each state
using separate data frames}
\usage{
diagnose_transitions(g)
}
\description{
Take a look at the computed transitions for each state
using separate data frames
}
