% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staggered.R
\name{eq_combine_xgroup}
\alias{eq_combine_xgroup}
\title{Aggregate equilibrium behavior in games with random active player}
\usage{
eq_combine_xgroup(g, eq = g[["eq"]], ap.col = ifelse(has.col(eq, "ap"),
  "ap", NA))
}
\arguments{
\item{g}{the game object}

\item{eq}{the equilibrium, by default the last solved eq of g.}

\item{ap.col}{the column as a character in x.df that is the index of the active player. By default "ap".}
}
\description{
Often it is useful to specify games such that players don't
move simultaneously but a random player ap is chosen to be active
in a given state.
}
\details{
The active player in a state x is defined by the variable ap
in x.df and the original state by xgroup.

This function aggregates equilibrium outcomes from x to xgroup.
For payoffs r1,r2,v1,v2 and U we take the mean over the payoffs
given the two possible actvive players.

The columns move.adv1 and move.adv2 describe the difference in
negotiation payoffs of a player when is the active player who can
make a move compared to the other player being active.

Finally we create action labels by combining the actions
chosen when a player is active.
}
