% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rne_capped.R
\name{get_T_rne_history}
\alias{get_T_rne_history}
\title{Get the intermediate steps in from t = T to t = 1 for
a T-RNE or capped RNE that has been solved with
\code{save.history = TRUE}}
\usage{
get_T_rne_history(g)
}
\description{
Get the intermediate steps in from t = T to t = 1 for
a T-RNE or capped RNE that has been solved with
\code{save.history = TRUE}
}
