% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transititions.R
\name{independent_transitions}
\alias{independent_transitions}
\title{Helper function to specify state transitions}
\usage{
independent_transitions(ax.df, ..., enclos = parent.frame(),
  remove.zero.prob = TRUE, prob.var = "prob")
}
\description{
See vignette for examples
}
