% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irv.R
\name{irv}
\alias{irv}
\title{Helper functions to specify state transitions}
\usage{
irv(var, ..., default = NULL, lower = NULL, upper = NULL,
  vals.unique = TRUE)
}
\description{
To be used as argument of \code{\link{irv_joint_dist}}
}
\details{
See vignette for examples
}
