% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irv.R
\name{irv_joint_dist}
\alias{irv_joint_dist}
\title{Helper function to specify state transitions}
\usage{
irv_joint_dist(
  df,
  ...,
  enclos = parent.frame(),
  remove.zero.prob = TRUE,
  prob.var = "prob"
)
}
\description{
See vignette for examples
}
