% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_eq_payoff_set}
\alias{plot_eq_payoff_set}
\title{Show a base R plot of equilibrium payoff set}
\usage{
plot_eq_payoff_set(
  g,
  x = eq$x[1],
  t = 1,
  eq = if (use.vr) get_eq(g, add.vr = TRUE) else g[["eq"]],
  xlim = NULL,
  ylim = NULL,
  add = FALSE,
  plot.r = TRUE,
  alpha = 0.8,
  black.border = TRUE,
  add.state.label = is.null(labels),
  labels = NULL,
  colors = c("#377EB8", "#E41A1C", "#4DAF4A", "#984EA3", "#FF7F00", "#FFFF33",
    "#A65628", "#F781BF"),
  add.xlim = NULL,
  add.ylim = NULL,
  extend.lim.perc = 0.05,
  use.vr = FALSE,
  ...
)
}
\arguments{
\item{g}{The game object for which an equilibrium has been solved}

\item{x}{A character vector of the state(s) for which the (continuation) equilibrium payoff set shall be shown. By default only the first stage.}

\item{eq}{An equilibrium object. By default the last solved equilibrium.}

\item{xlim}{as in \code{\link{plot.default}}}

\item{ylim}{as in \code{\link{plot.default}}}

\item{add}{as in \code{\link{plot.default}} Setting \code{add=FALSE} can be useful to compare payoff sets of different games.}

\item{plot.r}{Shall negotiation payoffs be shown as a point on the Pareto-frontier (default = TRUE)}

\item{alpha}{opacity of the fill color}
}
\description{
Show a base R plot of equilibrium payoff set
}
