% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistage.R
\name{rel.capped.rne.multistage.old}
\alias{rel.capped.rne.multistage.old}
\title{Solve an RNE for a capped version of a multistage game}
\usage{
rel.capped.rne.multistage.old(g, T, save.details = FALSE, tol = 1e-10,
  delta = g$param$delta, rho = g$param$rho, res.field = "eq",
  tie.breaking = c("slack", "random", "first", "last")[1], add = TRUE,
  keep.all.t = FALSE)
}
\description{
Solve an RNE for a capped version of a multistage game
}
