% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rne_capped.R
\name{rel_capped_rne_old}
\alias{rel_capped_rne_old}
\title{Solve an RNE for a capped version of the game}
\usage{
rel_capped_rne_old(g, T, save.details = FALSE, tol = 1e-10,
  delta = g$param$delta, rho = g$param$rho, adjusted.delta = NULL,
  res.field = "rne", tie.breaking = c("slack", "random", "first",
  "last")[1], use.cpp = TRUE)
}
\arguments{
\item{g}{The game object}

\item{T}{The number of periods until states can change}

\item{save.details}{If yes, detailed information about the equilibrium for each state and period will be stored in g and can be retrieved via the function get.rne.details}
}
\description{
Once the capped version of the game reaches period T,
the state cannot change anymore.
We can solve such capped games via a fast backward induction
algorithm. There always exists a unique RNE payoff.
}
