% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spe_classic.R
\name{rel_capped_spe}
\alias{rel_capped_spe}
\title{Solve for the set of SPE payoffs in every state and for optimal simple equilibria}
\usage{
rel_capped_spe(g, T, delta = g$param$delta, res.field = "eq", ...)
}
\description{
This function is just an interface to the dyngame package
}
