% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rne.R
\name{rel_eq_as_discounted_sums}
\alias{rel_eq_as_discounted_sums}
\title{Translate equilibrium payoffs as discounted sum
of payoffs}
\usage{
rel_eq_as_discounted_sums(g)
}
\description{
By default equilibrium payoffs are given as average discounted payoffs. This is the discounted sum of payoffs multiplied by (1-delta).
}
\details{
Call this function after you have solved an equilibrium if you want to present the equilibrium

@param g the game for which an equilibrium was computedayoffs as the discounted sum of payoffs instead.
}
