% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spe.R
\name{rel_first_best}
\alias{rel_first_best}
\title{Compute first-best.}
\usage{
rel_first_best(g, delta = g$param$delta, ...)
}
\arguments{
\item{g}{the game object}

\item{delta}{The discount factor}

\item{...}{additional parameters of \code{\link{rel_spe}}}
}
\description{
We compute the "equilibrium" play that would maximize
joint payoffs if incentive constraints could be
completely ignored.
}
\details{
Note that we create the same columns as for a spe, e.g.
punishment payoffs v1 and v2 that would arise if every
action profile could be implemented as punishment.
This allows to use the same functions, like \code{eq_diagram} as for equilibria.
}
