% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rne_from_actions.R
\name{rel_is_eq_rne}
\alias{rel_is_eq_rne}
\title{Checks if an equilibrium eq with negotiation payoffs is an RNE}
\usage{
rel_is_eq_rne(g, eq = g[["eq"]], r1 = eq$r1, r2 = eq$r2, r.tol = 1e-10,
  verbose = TRUE)
}
\description{
We simply solve the truncated game with r1 and r2
and check whether the resultig r1 and r2 are the same
}
