% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relhold.R
\name{rel_param}
\alias{rel_param}
\title{Add parameters to a relational contracting game}
\usage{
rel_param(g, ..., delta = non.null(param[["delta"]], 0.9),
  rho = non.null(param[["rho"]], 0), beta1 = non.null(param[["beta1"]],
  1/2), param = g[["param"]])
}
\arguments{
\item{g}{a relational contracting game created with rel_game}

\item{...}{other parameters that can e.g. be used in payoff functions}

\item{delta}{The discount factor}

\item{rho}{The negotiation probability}
}
\value{
Returns the updated game
}
\description{
Add parameters to a relational contracting game
}
