% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relhold.R
\name{rel_payoff}
\alias{rel_payoff}
\title{Add a payoff function to one or several states}
\usage{
rel_payoff(g, x = NULL, pi1 = 0, pi2 = 0)
}
\arguments{
\item{g}{a relational contracting game created with rel_game}

\item{x}{Name of one or multiple states. If NULL assume the payoff function holds for all states}

\item{pi1}{Player 1's payoff. Value(s) or formula}

\item{pi2}{Player 2's payoff. Values(s) or formula}
}
\value{
Returns the updated game
}
\description{
Add a payoff function to one or several states
}
