% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relhold.R
\name{rel_payoff_fun}
\alias{rel_payoff_fun}
\title{Version of rel_payoff that takes a function}
\usage{
rel_payoff_fun(g, pi.fun = NULL, x = NULL, vec.pi.fun = NULL, ...)
}
\description{
Version of rel_payoff that takes a function
}
