% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rne.R
\name{rel_rne}
\alias{rel_rne}
\title{Find an RNE for a (weakly) directional game}
\usage{
rel_rne(g, delta = g$param$delta, rho = g$param$rho,
  adjusted.delta = NULL, beta1 = g$param$beta1, verbose = TRUE, ...)
}
\arguments{
\item{g}{The game object}
}
\description{
If the game is strongly directional,
i.e. non-terminal states will be reached at most once,
there exists a unique RNE payoff.
}
\details{
For weakly directional games no RNE or multiple
RNE payoffs may exist.

You can call rel_capped_rne to solve a capped version of the game that allows state changes only up to some period T. Such a capped version always has a unique RNE payoff.
}
