% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rne.R
\name{rel_scale_eq_payoffs}
\alias{rel_scale_eq_payoffs}
\title{Scale equilibrium payoffs}
\usage{
rel_scale_eq_payoffs(g, factor)
}
\arguments{
\item{g}{the game for which an equilibrium was computed}

\item{factor}{the factor by which the payoffs U,v1,v2,r1 and r2 are multiplied}
}
\description{
Scale equilibrium payoffs
}
