% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spe_trunc.R
\name{rel_spe}
\alias{rel_spe}
\title{Finds an optimal simple subgame perfect equilibrium of g. From this the whole SPE payoff set can be deduced.}
\usage{
rel_spe(g, delta = g$param$delta, rho = g$param$rho, tol.feasible = 1e-10,
  verbose = FALSE, r1 = NULL, r2 = NULL, add.action.labels = TRUE,
  max.iter = 10000, no.exist.action = c("warn", "stop", "nothing"))
}
\arguments{
\item{g}{the considered game}

\item{r1}{(or \code{r2}) if not NULL we want to find a SPE in a truncated game. Then r1 and r2 need to specify for each state the exogenously fixed negotiation payoffs.}
}
\description{
Finds an optimal simple subgame perfect equilibrium of g. From this the whole SPE payoff set can be deduced.
}
