% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relhold.R
\name{rel_state}
\alias{rel_state}
\title{Add a state to a relational contracting game}
\usage{
rel_state(g, x, A1 = list(a1 = ""), A2 = list(a2 = ""), pi1 = NULL,
  pi2 = NULL, x.T = NULL)
}
\arguments{
\item{g}{a relational contracting game created with rel_game}

\item{x}{The name of the state}

\item{A1}{The action set of player 1. Can be a numeric or character vector (possible a formula creating one).}

\item{A2}{The action set of player 2. Can be a numeric or character vector  (possible a formula creating one).}

\item{pi1}{Player 1's payoff. Value(s) or formula}

\item{pi2}{Player 2's payoff. Values(s) or formula}

\item{x.T}{Relevant when solving a capped game. Which terminal state shall be set in period T onwards. By default, we stay in state x.}
}
\value{
Returns the updated game
}
\description{
Add a state to a relational contracting game
}
