% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relhold.R
\name{rel_transition}
\alias{rel_transition}
\title{Add a state transition from one state to one or several states}
\usage{
rel_transition(g, xs, xd, ..., prob = 1)
}
\arguments{
\item{g}{a relational contracting game created with rel_game}

\item{xs}{Name(s) of source states}

\item{xd}{Name(s) of destination states}

\item{...}{named action and their values}

\item{prob}{transition probability}
}
\value{
Returns the updated game
}
\description{
Add a state transition from one state to one or several states
}
