% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relhold.R
\name{rel_transition_fun}
\alias{rel_transition_fun}
\title{Add a state transition function}
\usage{
rel_transition_fun(g, trans.fun, x = NULL, ..., vectorized = FALSE)
}
\arguments{
\item{g}{a relational contracting game created with rel_game}

\item{...}{named action and their values}

\item{xs}{Name(s) of source states}

\item{xd}{Name(s) of destination states}

\item{prob}{transition probability}
}
\value{
Returns the updated game
}
\description{
Add a state transition function
}
