% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transititions.R
\name{trans_val}
\alias{trans_val}
\title{Helper functions to specify state transitions}
\usage{
trans_val(val, prob)
}
\description{
To be used as argument of \code{\link{trans_var}}
}
\details{
See vignette for examples
}
