% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transititions.R
\name{trans_var}
\alias{trans_var}
\title{Helper functions to specify state transitions}
\usage{
trans_var(var, ..., default = NULL, lower = NULL, upper = NULL,
  vals.unique = TRUE)
}
\description{
To be used as argument of \code{\link{independent_transitions}}
}
\details{
See vignette for examples
}
