% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spe_trunc.R
\name{trunc.spe.highest.U}
\alias{trunc.spe.highest.U}
\title{Calculates the highest joint payoff
The returned policy are indexed on ax (not on admiss)}
\usage{
\method{trunc}{spe.highest.U}(g, admiss, admiss.sizes, r1 = g[["r1"]],
  r2 = g[["r2"]], is.multi.stage = isTRUE(g$is.multi.stage),
  static.Pi = NULL)
}
\description{
Calculates the highest joint payoff
The returned policy are indexed on ax (not on admiss)
}
