% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rne_capped.R
\name{rel_capped_rne}
\alias{rel_capped_rne}
\title{Solve an RNE for a capped version of a multistage game}
\usage{
rel_capped_rne(g, T, tol = 1e-10, delta = g$param$delta,
  rho = g$param$rho, adjusted.delta = NULL, res.field = "eq",
  tie.breaking = c("equal_r", "slack", "random", "first", "last", "max_r1",
  "max_r2")[1], use.cpp = TRUE, save.details = FALSE,
  add.iterations = FALSE, save.history = FALSE, add.stationary = FALSE,
  T.rne = FALSE, spe = NULL)
}
\description{
Solve an RNE for a capped version of a multistage game
}
