% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rne.R
\name{rel_rne}
\alias{rel_rne}
\title{Find an RNE for a (weakly) directional game}
\usage{
rel_rne(
  g,
  delta = g$param$delta,
  rho = g$param$rho,
  adjusted.delta = NULL,
  beta1 = g$param$beta1,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{g}{The game object}

\item{delta}{the discount factor}

\item{rho}{the negotiation probability}

\item{adjusted.delta}{the adjusted discount factor (1-rho)*delta. Can be specified instead of delta.}

\item{beta1}{the bargaining weight of player 1. By default equal to 0.5. Can also be initially specified with \code{rel_param}.}

\item{verbose}{if \code{TRUE} give more detailed information over the solution process.}
}
\description{
If the game is strongly directional,
i.e. non-terminal states will be reached at most once,
there exists a unique RNE payoff.
}
\details{
For weakly directional games no RNE or multiple
RNE payoffs may exist.

You can call rel_capped_rne to solve a capped version of the game that allows state changes only up to some period T. Such a capped version always has a unique RNE payoff.
}
