% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rne_capped.R
\name{rel_T_rne}
\alias{rel_T_rne}
\title{Compute a T-RNE}
\usage{
rel_T_rne(g, T, tol = 1e-10, delta = g$param$delta, rho = g$param$rho,
  adjusted.delta = NULL, res.field = "eq", tie.breaking = c("equal_r",
  "slack", "random", "first", "last", "max_r1", "max_r2")[1], use.cpp = TRUE,
  save.details = FALSE, add.iterations = FALSE, save.history = FALSE,
  add.stationary = FALSE, T.rne = TRUE, spe = g[["spe"]])
}
\arguments{
\item{g}{The game}

\item{T}{The number of periods in which new negotiations can take place.}
}
\description{
The idea of a T-RNE is that only for a finite number of T periods relational contracts will be newly negoatiated. After T periods no new negotiations take place, i.e. every SPE continuation payoff can be implemented. For fixed T there is a unique RNE payoff.
}
